fix(trade): 修复绑定分销员 ID 的类型兼容性
- 引入 FlexInt64 类型以支持 JSON 中的数字与字符串格式互转
- 解决前端可能传递字符串类型 ID 导致绑定失败的问题

+++ internal/api/handler/app/mall/trade/brokerage/user.go
@@ -67,7 +67,7 @@ func (h *AppBrokerageUserHandler) BindBrokerageUser(c *gin.Context) {
 		return
 	}
 	userId := c.GetInt64("userId")
-	success, err := h.userSvc.BindBrokerageUser(c, userId, r.BindUserID)
+	success, err := h.userSvc.BindBrokerageUser(c, userId, int64(r.BindUserID))
 	if err != nil {
 		response.WriteError(c, 500, err.Error()) // Or 400
 		return
